move the function to get Spotify API access token to home page file
